+++ src/Servebolt/CachePurge/WordPressCachePurge/TermMethods.php
@@ -2,8 +2,11 @@
 
 namespace Servebolt\Optimizer\CachePurge\WordPressCachePurge;
 
+if (!defined('ABSPATH')) exit; // Exit if accessed directly
+
 use Servebolt\Optimizer\CachePurge\CachePurge as CachePurgeDriver;
 use Servebolt\Optimizer\CachePurge\PurgeObject\PurgeObject;
+use Servebolt\Optimizer\Queue\Queues\WpObjectQueue;
 
 /**
  * Trait TermMethods
@@ -38,9 +41,17 @@ trait TermMethods
      */
     public static function purgeTermCache(int $termId, string $taxonomySlug): bool
     {
-        // TODO: Add queue handling here
-        $urlsToPurge = self::getUrlsToPurgeByTermId($termId, $taxonomySlug);
-        $cachePurgeDriver = CachePurgeDriver::getInstance();
-        return $cachePurgeDriver->purgeByUrls($urlsToPurge);
+        if (CachePurgeDriver::queueBasedCachePurgeIsActive()) {
+            $queueInstance = WpObjectQueue::getInstance();
+            return $queueInstance->add([
+                $termId,
+                'term',
+                ['taxonomy_slug' => $taxonomySlug]
+            ]);
+        } else {
+            $urlsToPurge = self::getUrlsToPurgeByTermId($termId, $taxonomySlug);
+            $cachePurgeDriver = CachePurgeDriver::getInstance();
+            return $cachePurgeDriver->purgeByUrls($urlsToPurge);
+        }
     }
 }
